# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Groene Vingers",
        "version": "3.0.0"
    },
    "paths": {
...
        "/api/product/checkoutproduct/all": {
            "get": {
                "tags": [
                    "Checkout"
                ],
                "summary": "Show list of all checkoutproducts",
                "description": "Show list of all checkoutproducts",
                "operationId": "checkoutProduct_All",
                "responses": {
                    "200": {
                        "description": "Show list of all checkoutproducts"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api_gv/public/api/product/code/{code}": {
            "get": {
                "tags": [
                    "Checkout"
                ],
                "summary": "Show product info by code",
                "description": "Show product info by code",
                "operationId": "show_product_by_code",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "code, eg; 123456789",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show product info by code"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api_gv/public/api/product/checkoutproduct": {
            "post": {
                "tags": [
                    "Checkout"
                ],
                "summary": "Post product to table",
                "description": "Post product to table",
                "operationId": "post_product",
                "requestBody": {
                    "description": "Post product to table",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "article_number",
                                    "name",
                                    "price",
                                    "user_id"
                                ],
                                "properties": {
                                    "article_number": {
                                        "type": "integer",
                                        "example": "123456789"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "product name"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": "123456789"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": "123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post product to table"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Checkout"
                ],
                "summary": "Delete all checkoutproducts",
                "description": "Delete all checkoutproducts",
                "operationId": "delete_all_checkoutproducts",
                "responses": {
                    "200": {
                        "description": "Delete all checkoutproducts"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api_gv/public/api/product/barcode/{barcode}": {
            "post": {
                "tags": [
                    "Checkout"
                ],
                "summary": "Post product to other table by searching barcode",
                "description": "Post product to other table by searching barcode",
                "operationId": "post_product_by_barcode",
                "parameters": [
                    {
                        "name": "barcode",
                        "in": "path",
                        "description": "barcode, eg; 123456789",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post product to other table by searching barcode"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api_gv/public/api/product/checkoutproduct/{id}": {
            "delete": {
                "tags": [
                    "Checkout"
                ],
                "summary": "Delete checkoutproduct by id",
                "description": "Delete checkoutproduct by id",
                "operationId": "delete_checkoutproduct_by_id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete checkoutproduct by id"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api_gv/public/api/product/receipt": {
            "post": {
                "tags": [
                    "Checkout"
                ],
                "summary": "receipt_product to receipt",
                "description": "receipt_product to receipt",
                "operationId": "post_receipt",
                "responses": {
                    "200": {
                        "description": "receipt_product to receipt"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api_gv/public/api/receipts": {
            "get": {
                "tags": [
                    "Checkout"
                ],
                "summary": "Get all receipts",
                "description": "Get all receipts",
                "operationId": "get_all_receipts",
                "responses": {
                    "200": {
                        "description": "Get all receipts"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api_gv/public/api/login": {
            "post": {
                "tags": [
                    "Checkout"
                ],
                "summary": "Login with customernumber",
                "description": "Login with customernumber",
                "operationId": "login_with_customernumber",
                "requestBody": {
                    "description": "Login with customernumber",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customernumber",
                                    "password"
                                ],
                                "properties": {
                                    "customernumber": {
                                        "type": "integer",
                                        "example": "123456789"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login with customernumber"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api_gv/public/api/user/customernumber/{customernumber}": {
            "get": {
                "tags": [
                    "Checkout"
                ],
                "summary": "Get user by customernumber",
                "description": "Get user by customernumber",
                "operationId": "user_by_customernumber",
                "parameters": [
                    {
                        "name": "customernumber",
                        "in": "path",
                        "description": "customernumber, eg; 123456789",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get user by customernumber"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api_gv/public/api/user/customernumber/all": {
            "get": {
                "tags": [
                    "Checkout"
                ],
                "summary": "Get all customers",
                "description": "Get all customers",
                "operationId": "getAllCustomers",
                "responses": {
                    "200": {
                        "description": "Get all customers"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Bearer token, Ex: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearer": {
                "type": "apiKey",
                "description": "Bearer token, Ex: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    ]
}